requirements.txt
websocket-client
requests
tls-client
pypasser
names
colorama
curl_cffi
streamlit
selenium
fake-useragent
twocaptcha
https://github.com/AI-Yash/st-chat/archive/refs/pull/24/head.zip
pydantic
pymailtm
Levenshtein
retrying
mailgw_temporary_email
pycryptodome
random-password-generator
numpy>=1.13.3
tornado>=6.3.2
plotly==5.14.1
matplotlib==3.4.2
seaborn==0.11.1
scikit-learn==0.24.2
tensorflow==2.5.0
keras==2.4.3
statsmodels==0.12.2
scipy==1.7.0
jupyterlab==3.1.10
fastai==2.5.2
xgboost==1.4.2
lightgbm==3.2.1
catboost==0.26.1
nltk==3.6.2
gensim==4.0.1
pytorch-lightning==1.4.8
python-dotenv==1.0.0
replicate
google-generativeai